requirements.txt
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
defusedxml==0.5.0
Django==1.10.4
-e git+https://github.com/zmrenwu/django-activity-stream.git@c0b0448c2941f7cb350ed4b97ffe9eeeb3bc456e#egg=django_activity_stream
django-ajax==0.2.0
django-allauth==0.29.0
django-anymail==0.10
django-appconf==1.0.2
django-autoslug==1.9.3
django-bootstrap-pagination==1.6.2
django-bootstrap3==9.1.0
django-braces==1.10.0
django-ckeditor==5.4.0
django-contrib-comments==1.7.3
django-crispy-forms==1.6.1
django-haystack==2.6.1
django-imagekit==3.3
django-js-asset==1.0.0
django-model-utils==2.6
django-mptt==0.8.6
django-notifications-hq==1.2
django-pagedown==0.1.1
django-robots==3.0
django-simple-captcha==0.5.3
django-taggit==0.21.3
djangoajax==2.4
Flask==0.12.2
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
jsonfield==2.0.2
Markdown==2.6.7
MarkupSafe==1.0
oauthlib==2.0.4
pilkit==2.0
Pillow==3.4.2
Pygments==2.1.3
python-slugify==1.2.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
timeago==1.0.7
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
Whoosh==2.7.4
